MidiGenerator - move simple static methods to MidiGeneratorUtils
